+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/manager/DeveloperManager.java
@@ -6,8 +6,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 
 import gov.healthit.chpl.dao.EntityCreationException;
 import gov.healthit.chpl.dao.EntityRetrievalException;
-import gov.healthit.chpl.dto.DecertifiedDeveloperDTO;
 import gov.healthit.chpl.dto.DeveloperDTO;
+import gov.healthit.chpl.web.controller.results.DecertifiedDeveloperResults;
 
 public interface DeveloperManager {
 	public List<DeveloperDTO> getAll();
@@ -19,5 +19,5 @@ public interface DeveloperManager {
 	public void delete(Long developerId) throws EntityRetrievalException, JsonProcessingException, EntityCreationException;
 	public DeveloperDTO merge(List<Long> developerIdsToMerge, DeveloperDTO developerToCreate) throws EntityRetrievalException, JsonProcessingException, EntityCreationException;
 	public void checkSuspiciousActivity(DeveloperDTO original, DeveloperDTO changed);
-	public List<DecertifiedDeveloperDTO> getDecertifiedDevelopers();
+	public DecertifiedDeveloperResults getDecertifiedDevelopers() throws EntityRetrievalException;
 }