+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/web/controller/results/AnnouncementResults.java
@@ -1,12 +1,13 @@
 package gov.healthit.chpl.web.controller.results;
 
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
 import gov.healthit.chpl.domain.Announcement;
 
-public class AnnouncementResults {
-	
+public class AnnouncementResults implements Serializable {
+	private static final long serialVersionUID = 8748247550112564530L;
 	private List<Announcement> announcements;
 
 	public AnnouncementResults() {