+++ RELEASE_NOTES.md
@@ -10,6 +10,8 @@ _Date TBD_
   * Add caching for /certification_ids API call
   * Add caching for /pending API call
   * Add caching for /search API call
+  * Add caching for /decertifications/developers
+  * Add caching for /decertifications/certified_products
   * Update caching for /search_options API call
   * Cache underlying data that improves call speed for many other API calls
 * TO-DO DURING DEPLOYMENT: Add these lines to environment.properties: